requirements.txt
mysql-connector-python==8.0.28
requests==2.27.1
dateparser==1.1.0
GitPython==3.1.26